requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6

# HTML parsing
beautifulsoup4==4.12.3

# Database and ORM
SQLAlchemy==2.0.35
psycopg[binary]>=3.1.19
alembic==1.13.2
pgvector>=0.2.0

# Configuration and validation
pydantic-settings==2.4.0
python-dotenv==1.0.1

# Authentication and JWT
PyJWT>=2.8.0

# LLM providers
openai>=1.0.0
anthropic>=0.25.0
pyyaml>=6.0

# Redis / Queue (hiredis provides faster C parser for production)
redis[hiredis]>=5.0.8

# Observability and Monitoring
prometheus-client==0.20.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-requests==0.47b0

# Logging
loguru==0.7.2

# AI and ML
openai>=1.3.0
anthropic>=0.7.0
chromadb>=0.4.15
tiktoken>=0.5.0  # Token counting for embeddings

# Vector database and embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# Async HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# Git operations
GitPython>=3.1.40

# Monitoring
structlog>=23.2.0

# Additional utilities
python-multipart>=0.0.6

# Task queue
dramatiq==1.16.0

# Integrations dependencies  
httpx==0.27.2
pydantic==2.9.2
python-json-logger==2.0.7

# Security - Token encryption
cryptography>=41.0.0
boto3>=1.28.0
# JWT authentication
python-jose[cryptography]>=3.3.0

# Optional: Legacy scripts (backfill_pgvector.py uses psycopg2)
psycopg2-binary>=2.9.0
boto3>=1.28.0
slack_sdk==3.27.1  # Slack Web API client
msal==1.28.0  # Microsoft Authentication Library for Graph API
